chore: Move ParticipantsConversationSystemMessageCell to its own file
